requirements.txt
beautifulsoup4==4.12.2
nltk==3.8.1
oauthlib==3.2.2
openai==0.27.5
pandas==2.0.1
keras==2.12.0
scikit-learn==1.2.2
selenium==4.9.0
spacy==3.5.2
swifter==1.3.4
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
tqdm==4.65.0
webdriver-manager==3.8.6
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.5.0/pt_core_news_lg-3.5.0-py3-none-any.whl